refactor: primitive internals
