+++ geometric/tests/test_diff.py
@@ -8,6 +8,7 @@ import json, os, shutil
 
 import geometric.coordinate_systems.cartesian
 import geometric.coordinate_systems.delocalised
+import geometric.coordinate_systems.primitive
 from . import addons
 import geometric
 import pytest
@@ -19,7 +20,7 @@ datad = addons.datad
 def test_diff_h2o2_h2o(localizer):
     M = geometric.molecule.Molecule(os.path.join(datad, 'h2o2_h2o.pdb'))
     IC = geometric.coordinate_systems.delocalised.DelocalizedInternalCoordinates(M, build=True, connect=False, addcart=False)
-    IC1 = geometric.internal.PrimitiveInternalCoordinates(M, build=True, connect=False, addcart=False)
+    IC1 = geometric.coordinate_systems.primitive.PrimitiveInternalCoordinates(M, build=True, connect=False, addcart=False)
     IC2 = geometric.coordinate_systems.cartesian.CartesianCoordinates(M)
     IC3 = geometric.coordinate_systems.delocalised.DelocalizedInternalCoordinates(M, build=True, connect=True, addcart=False)
     assert IC1.repr_diff(IC) == "Primitive -> Delocalized"
